+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `admin-reset-reviews` command to reset stored reviews
+- Add optional merge strategy override to `bot merge` command
 
 ### Fixed
 
 - Remove reviews before removing pull request (especially needed on `bot admin-disable`)
+- Handle empty body on issue/pull request
 
 ### Changed